cp_headers: Add legacy Elite audio framework headers
Add the headers supplied by the legacy Elite audio
framework. These are required for compiling the legacy
Audio HAL as well as some BSP components, such as
ACDB-LOADER. In order for the headers of the Audioreach
and Elite audio frameworks to coexist, the latter were
placed in the legacy directory.

The list of AudioReach headers was also synched with what
is supplied by the corresponding techpack-audio package.

Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>